+++ NEWS.md
@@ -9,7 +9,10 @@
 
  * `futurize()` gained argument `eval`, which can be used to return
    the futurized expression instead of evaluating it.
-   
+
+ * Add support for `times()` of **foreach**. It requires special care
+   for passing future options, e.g. `futurize(seed = TRUE)`.
+ 
 
 # Version 0.0.2 (2025-05-23)